+++ src/app/in-viewport/in-viewport.module.ts
@@ -1,8 +1,8 @@
 import { NgModule, ModuleWithProviders } from '@angular/core';
 
 import { InViewportDirective } from './in-viewport.directive';
-import { WINDOW } from '../window/window-token';
-import { WINDOW_MOCK } from '../window/window-mock';
+import { WINDOW } from '../..';
+import { WINDOW_MOCK } from '../..';
 
 /**
  * A simple lightweight library for Angular with other dependencies
@@ -11,13 +11,13 @@ import { WINDOW_MOCK } from '../window/window-mock';
  */
 @NgModule({
   declarations: [InViewportDirective],
-  exports: [InViewportDirective]
+  exports: [InViewportDirective],
 })
 export class InViewportModule {
   static forServer(): ModuleWithProviders {
     return {
       ngModule: InViewportModule,
-      providers: [{ provide: WINDOW, useValue: WINDOW_MOCK }]
+      providers: [{ provide: WINDOW, useValue: WINDOW_MOCK }],
     };
   }
 }