+++ webpack.server.js
@@ -22,7 +22,7 @@ module.exports = {
       {
         // Mark files inside `@angular/core` as using SystemJS style dynamic imports.
         // Removing this will cause deprecation warnings to appear.
-        test: /(\\|\/)@angular(\\|\/)core(\\|\/).+\.js$/,
+        test: /([\\/])@angular([\\/])core([\\/]).+\.js$/,
         parser: { system: true }
       }
     ]
@@ -30,13 +30,13 @@ module.exports = {
   plugins: [
     new webpack.ContextReplacementPlugin(
       // fixes WARNING Critical dependency: the request of a dependency is an expression
-      /(.+)?angular(\\|\/)core(.+)?/,
+      /(.+)?angular([\\/])core(.+)?/,
       path.join(__dirname, 'src'), // location of your src
       {} // a map of your routes
     ),
     new webpack.ContextReplacementPlugin(
       // fixes WARNING Critical dependency: the request of a dependency is an expression
-      /(.+)?express(\\|\/)(.+)?/,
+      /(.+)?express([\\/])(.+)?/,
       path.join(__dirname, 'src'),
       {}
     )